fix: broken/jumbled tests
